refactor(libcore): archive JNI api husk, relocate conn_stats atomics
CONNECTION_STATE + CONNECTION_START_TIME move from the doomed
api/conn_stats.rs to state.rs (their only external readers — events and
quic/server — now import from there). The rest of the hand-rolled JNI
surface (initApi, connect, the useless network stats, getPublicAddr, the
live identity-pairing server, welcome, computeQrMask) is archived to
../.archive/libcore-jni-api/. api/ is now an empty leaf stub awaiting the
uniffi translation layer. ndk/ is dead code now; swept once SecureStore
lands.
